docs : move files to each folders
- add : Common, ListPage, LoginPage, Search folders in src/components

- move files into each folders

- change import file routes
